requirements/requirements.txt
django-ckeditor==5.4.0
django-mptt==0.9.0
django-notifications-hq==1.3
Pillow==5.0.0
